修复文章评论模块bug以及添加测试数据

+++ src/main/java/com/zhlzzz/together/controllers/article/AdvertController.java
@@ -4,7 +4,10 @@ import com.zhlzzz.together.article.advert.Advert;
 import com.zhlzzz.together.article.advert.AdvertEntity;
 import com.zhlzzz.together.article.advert.AdvertParam;
 import com.zhlzzz.together.article.advert.AdvertService;
+import com.zhlzzz.together.controllers.ApiAuthentication;
 import com.zhlzzz.together.controllers.ApiExceptions;
+import com.zhlzzz.together.user.User;
+import com.zhlzzz.together.user.UserService;
 import com.zhlzzz.together.utils.CollectionUtils;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -19,7 +22,6 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @RestController
@@ -30,6 +32,7 @@ import java.util.stream.Collectors;
 public class AdvertController {
 
     private final AdvertService advertService;
+    private final UserService userService;
 
     @GetMapping
     @ApiOperation(value = "获取广告列表")
@@ -43,10 +46,14 @@ public class AdvertController {
     @PostMapping
     @ApiOperation(value = "新增广告")
     @ResponseBody
-    public AdvertView addAdvert(@Valid @RequestBody AdvertParam advertParam, BindingResult result) {
+    public AdvertView addAdvert(@Valid @RequestBody AdvertParam advertParam, ApiAuthentication auth, BindingResult result) {
         if (result.hasErrors()) {
             String errors = result.getAllErrors().stream().map((e)->e.toString()).collect(Collectors.joining(";\n"));
             throw ApiExceptions.badRequest(errors);
+        }
+        User admin = userService.getUserById(auth.requireUserId()).filter(u -> u.isAdmin()).orElse(null);
+        if (admin == null) {
+            throw ApiExceptions.noPrivilege();
         }
         Advert advert = advertService.addAdvert(advertParam);
         return new AdvertView(advert);
@@ -55,10 +62,14 @@ public class AdvertController {
     @PutMapping(path = "/{id:\\d+}")
     @ApiOperation(value = "更新广告")
     @ResponseBody
-    public AdvertView updateAdvert(@PathVariable Long id, @Valid @RequestBody AdvertParam advertParam, BindingResult result) {
+    public AdvertView updateAdvert(@PathVariable Long id, @Valid @RequestBody AdvertParam advertParam, ApiAuthentication auth, BindingResult result) {
         if (result.hasErrors()) {
             String errors = result.getAllErrors().stream().map((e)->e.toString()).collect(Collectors.joining(";\n"));
             throw ApiExceptions.badRequest(errors);
+        }
+        User admin = userService.getUserById(auth.requireUserId()).filter(u -> u.isAdmin()).orElse(null);
+        if (admin == null) {
+            throw ApiExceptions.noPrivilege();
         }
         Advert advert = advertService.getAdvertById(id).orElseThrow(() -> ApiExceptions.notFound("不存在此广告"));
 
@@ -68,7 +79,11 @@ public class AdvertController {
     @DeleteMapping(path = "/{id:\\d+}")
     @ApiOperation(value = "删除广告")
     @ResponseBody
-    public ResponseEntity<String> delete(@PathVariable Long id) {
+    public ResponseEntity<String> delete(@PathVariable Long id,  ApiAuthentication auth) {
+        User admin = userService.getUserById(auth.requireUserId()).filter(u -> u.isAdmin()).orElse(null);
+        if (admin == null) {
+            throw ApiExceptions.noPrivilege();
+        }
         Advert advertEntity = advertService.getAdvertById(id).orElseThrow(() -> ApiExceptions.notFound("不存在此广告"));
         advertService.deleteAdvert(advertEntity.getId());
